requirements.txt
feedparser
requests
google-genai
newspaper3k